--- .firebase/hosting.YnVpbGQ.cache
@@ -1,10 +1,0 @@
-asset-manifest.json,1658908041301,2c4ac61bbdeed30fed4e7c40fd0b8c01cdb32500b11bf971355cc2b26c4bae59
-favicon.ico,1650662998143,eae62e993eb980ec8a25058c39d5a51feab118bd2100c4deebb2a9c158ec11f9
-brain.png,1658905588803,0a4df5ab0a6c67da6c102c920650207d6c65baebf0685d3e41edbac1e1fc1594
-index.html,1658908041300,edb78c4b2ffe879f92b7a80bcdc9dc12ed76992ac4e498b4d65d06ce48685d8a
-manifest.json,1650662998144,e5a5ca27c4444034354b69457d23d23d62d7c9040a66a4d4e402354afd824834
-static/js/main.ff134ec3.js.LICENSE.txt,1658908041305,344f60d92805a3f6078799efe3001c0746af425c166ecf1779b9a58f0ef566af
-static/css/main.fc7e42a3.css,1658908041306,79270d69384821ad0e435382f023e5ae1e3e15ecc8445751e01c869815ee21af
-static/css/main.fc7e42a3.css.map,1658908041306,c83a9a66a4789a09be5e8c2a133b7484f79a20e313ac4a267e59fe11c21f1b22
-static/js/main.ff134ec3.js,1658908041306,f7eb2e40bcf8d6742c6729e7e9ab9f258af9bf78448130f013a63f0163f0ad88
-static/js/main.ff134ec3.js.map,1658908041307,9c4e262a098ee771eb120a6b58eeef61704afcb3962874a17ae47182d2062993
