--- src/index.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-// import registerServiceWorker from './registerServiceWorker';
-import 'tachyons';
-
-ReactDOM.render(<App />, document.getElementById('root'));
-// registerServiceWorker();
